refactor for multiple user flows
