+++ packages/app/src/app/pages/Pro/pages/steps.ts
@@ -1,0 +1,5 @@
+export enum Steps {
+  WorkspacePlanSelection = 'WorkspacePlanSelection',
+  InlineCheckout = 'InlineCheckout',
+  ConfirmBillingInterval = 'ConfirmBillingInterval',
+}
